Gestione status code e validazione

+++ 20201019 - ripasso api/ITS.Dottor.IITS.Northwind/ITS.Dottor.IITS.Northwind.WebApi/Controllers/CategoriesController.cs
@@ -31,9 +31,13 @@ namespace ITS.Dottor.IITS.Northwind.WebApi.Controllers
 
         // GET api/Categories/5
         [HttpGet("{id}")]
-        public Category Get(int id)
+        public IActionResult Get(int id)
         {
-            return _categoriesRepository.GetById(id);
+            var category = _categoriesRepository.GetById(id);
+            if (category == null)
+                return NotFound();
+
+            return Ok(category);
         }
 
         // POST api/Categories
@@ -45,17 +49,27 @@ namespace ITS.Dottor.IITS.Northwind.WebApi.Controllers
 
         // PUT api/Categories/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Category category)
+        public IActionResult Put(int id, [FromBody] Category category)
         {
+            var dbCat = _categoriesRepository.GetById(id);
+            if (dbCat == null)
+                return NotFound();
+
             category.CategoryId = id;
             _categoriesRepository.Update(category);
+            return NoContent();
         }
 
         // DELETE api/Categories/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
+            var dbCat = _categoriesRepository.GetById(id);
+            if (dbCat == null)
+                return NotFound();
+
             _categoriesRepository.Delete(id);
+            return NoContent();
         }
     }
 }
